split away scopes and resolveToken
