+++ src/utils/parser/tests/fixtures/allSymbols.js
