+++ src/utils/parser/tests/fixtures/class.js
