+++ src/utils/parser/tests/fixtures/func.js
