+++ src/utils/parser/tests/fixtures/parseScriptTags.html
