+++ src/utils/parser/tests/fixtures/resolveToken.js
