+++ src/utils/parser/tests/fixtures/var.js
