+++ src/utils/parser/tests/getSymbols.js
@@ -1,21 +1,7 @@
 const expect = require("expect.js");
 import getSymbols from "../getSymbols";
 
-const fs = require("fs");
-const path = require("path");
-
-function getSourceText(name, type = "js") {
-  const text = fs.readFileSync(
-    path.join(__dirname, `../../tests/fixtures/${name}.${type}`),
-    "utf8"
-  );
-  const contentType = type === "html" ? "text/html" : "text/javascript";
-  return {
-    id: name,
-    text,
-    contentType
-  };
-}
+import { getSourceText } from "./helpers";
 
 describe("Parser.getSymbols", () => {
   describe("functions", () => {
