+++ src/utils/parser/tests/resolveToken.js
@@ -1,187 +1,10 @@
 const expect = require("expect.js");
-import {
-  getVariablesInLocalScope,
-  getVariablesInScope,
-  resolveToken
-} from "../parser/utils";
+import resolveToken from "../resolveToken";
+import { getClosestScope } from "../utils/closest";
 
-import { getClosestScope } from "../parser/utils/closest";
-
-const fs = require("fs");
-const path = require("path");
-
-function getSourceText(name, type = "js") {
-  const text = fs.readFileSync(
-    path.join(__dirname, `fixtures/${name}.${type}`),
-    "utf8"
-  );
-  const contentType = type === "html" ? "text/html" : "text/javascript";
-  return {
-    id: name,
-    text,
-    contentType
-  };
-}
+import { getSourceText } from "./helpers";
 
 describe("parser", () => {
-  describe("getVariablesInLocalScope", () => {
-    it("finds scope binding variables", () => {
-      const scope = getClosestScope(getSourceText("math"), {
-        line: 2,
-        column: 2
-      });
-
-      var vars = getVariablesInLocalScope(scope);
-      expect(vars.map(v => v.name)).to.eql(["n"]);
-      expect(vars[0].references[0].node.loc.start).to.eql({
-        column: 4,
-        line: 3
-      });
-    });
-
-    it("only gets local variables", () => {
-      const scope = getClosestScope(getSourceText("math"), {
-        line: 3,
-        column: 5
-      });
-
-      var vars = getVariablesInLocalScope(scope);
-
-      expect(vars.map(v => v.name)).to.eql(["n"]);
-      expect(vars[0].references[0].node.loc.start).to.eql({
-        column: 4,
-        line: 3
-      });
-    });
-
-    it("finds variables in block scope", () => {
-      const scope = getClosestScope(getSourceText("resolveToken"), {
-        line: 34,
-        column: 13
-      });
-
-      var vars = getVariablesInLocalScope(scope);
-
-      expect(vars.map(v => v.name)).to.eql(["x"]);
-    });
-  });
-
-  describe("getVariablesInScope", () => {
-    it("finds scope binding variables", () => {
-      const scope = getClosestScope(getSourceText("math"), {
-        line: 3,
-        column: 5
-      });
-
-      var vars = getVariablesInScope(scope);
-
-      expect(vars).to.eql([
-        "this",
-        "arguments",
-        "n",
-        "square",
-        "two",
-        "four",
-        "math",
-        "child"
-      ]);
-    });
-
-    it("finds variables from multiple scopes", () => {
-      let vars;
-      const source = getSourceText("resolveToken");
-
-      vars = getVariablesInScope(
-        getClosestScope(source, {
-          line: 36,
-          column: 19
-        })
-      );
-
-      expect(vars).to.eql([
-        "this",
-        "arguments",
-        "y",
-        "x",
-        "innerScope",
-        "outer",
-        "fromIIFE",
-        "a",
-        "b",
-        "getA",
-        "setB",
-        "plusAB",
-        "withMultipleScopes"
-      ]);
-
-      vars = getVariablesInScope(
-        getClosestScope(source, {
-          line: 34,
-          column: 14
-        })
-      );
-
-      expect(vars).to.eql([
-        "this",
-        "arguments",
-        "x",
-        "innerScope",
-        "outer",
-        "fromIIFE",
-        "a",
-        "b",
-        "getA",
-        "setB",
-        "plusAB",
-        "withMultipleScopes"
-      ]);
-
-      vars = getVariablesInScope(
-        getClosestScope(source, {
-          line: 24,
-          column: 9
-        })
-      );
-
-      expect(vars).to.eql([
-        "this",
-        "arguments",
-        "inner",
-        "innerScope",
-        "outer",
-        "fromIIFE",
-        "a",
-        "b",
-        "getA",
-        "setB",
-        "plusAB",
-        "withMultipleScopes"
-      ]);
-
-      vars = getVariablesInScope(
-        getClosestScope(source, {
-          line: 28,
-          column: 33
-        })
-      );
-
-      expect(vars).to.eql([
-        "this",
-        "arguments",
-        "toIIFE",
-        "innerScope",
-        "outer",
-        "fromIIFE",
-        "a",
-        "b",
-        "getA",
-        "setB",
-        "plusAB",
-        "withMultipleScopes"
-      ]);
-    });
-  });
-
   describe("resolveToken", () => {
     it("should get the expression for the token at location", () => {
       const { expression, inScope } = resolveToken(
