+++ src/utils/parser/types.js
@@ -1,0 +1,9 @@
+import type { SourceText, Location, Frame, TokenResolution } from "../../types";
+type Scope = {
+  location: {
+    line: number,
+    column: number
+  },
+  parent: Scope,
+  bindings: Object[]
+};
